moved checkin screens to their own folder
