+++ apps/passport-client/pages/index.tsx
@@ -12,11 +12,12 @@ import { HashRouter, Route, Routes } from "react-router-dom";
 import { AddScreen } from "../components/screens/AddScreen/AddScreen";
 import { AddSubscriptionScreen } from "../components/screens/AddSubscriptionScreen";
 import { ChangePasswordScreen } from "../components/screens/ChangePasswordScreen";
-import { DevconnectCheckinByIdScreen } from "../components/screens/DevconnectCheckinByIdScreen";
+import { DevconnectCheckinByIdScreen } from "../components/screens/Checkin/DevconnectCheckinByIdScreen";
+import { GenericIssuanceCheckInScreen } from "../components/screens/Checkin/PodboxScanScreen/GenericIssuanceCheckIn";
+import { SecondPartyTicketVerifyScreen } from "../components/screens/Checkin/SecondPartyTicketVerifyScreen";
 import { EnterConfirmationCodeScreen } from "../components/screens/EnterConfirmationCodeScreen";
 import { FrogManagerScreen } from "../components/screens/FrogScreens/FrogManagerScreen";
 import { FrogSubscriptionScreen } from "../components/screens/FrogScreens/FrogSubscriptionScreen";
-import { GenericIssuanceCheckInScreen } from "../components/screens/GenericIssuanceCheckIn";
 import { GetWithoutProvingScreen } from "../components/screens/GetWithoutProvingScreen";
 import { HaloScreen } from "../components/screens/HaloScreen/HaloScreen";
 import { HomeScreen } from "../components/screens/HomeScreen/HomeScreen";
@@ -32,7 +33,6 @@ import { MissingScreen } from "../components/screens/MissingScreen";
 import { NoWASMScreen } from "../components/screens/NoWASMScreen";
 import { ProveScreen } from "../components/screens/ProveScreen/ProveScreen";
 import { ScanScreen } from "../components/screens/ScanScreen";
-import { SecondPartyTicketVerifyScreen } from "../components/screens/SecondPartyTicketVerifyScreen";
 import { ServerErrorScreen } from "../components/screens/ServerErrorScreen";
 import { SubscriptionsScreen } from "../components/screens/SubscriptionsScreen";
 import { TermsScreen } from "../components/screens/TermsScreen";
